fixes #11322 - associate partition table and templates to Operating System upon creation
This commit is a minor change to the existing logic to ensure that when an Operating
System is created as part of a puppet checkin (e.g. the Katello server itself),
the resultant OS has the partition tables and provisioning templates associated.
Previously, this wasn't occuring because the OS created by foreman is using
Operatingsystem.create vs Redhat.create.  This makes sense as the server may
not be running Red Hat OS; howevever, it caused the redhat_extensions 'after_create'
filter to not get executed.
