+++ app/models/katello/concerns/redhat_extensions.rb
@@ -8,7 +8,6 @@ module Katello
         alias_method_chain :medium_uri, :content_uri
         alias_method_chain :mediumpath, :content
         alias_method_chain :boot_files_uri, :content
-        after_create :assign_templates!
       end
 
       module ClassMethods
@@ -43,22 +42,6 @@ module Katello
           else
             return family.gsub(' ', '_')
           end
-        end
-      end
-
-      def assign_templates!
-        # Automatically assign default templates
-        TemplateKind.all.each do |kind|
-          if (template = ProvisioningTemplate.find_by_name(Setting["katello_default_#{kind.name}"]))
-            provisioning_templates << template unless provisioning_templates.include?(template)
-            if OsDefaultTemplate.where(:template_kind_id => kind.id, :operatingsystem_id => id).empty?
-              OsDefaultTemplate.create(:template_kind_id => kind.id, :provisioning_template_id => template.id, :operatingsystem_id => id)
-            end
-          end
-        end
-
-        if (ptable = Ptable.find_by_name(Setting["katello_default_ptable"]))
-          ptables << ptable unless ptables.include?(ptable)
         end
       end
