+++ test/models/concerns/redhat_extensions_test.rb
@@ -38,25 +38,5 @@ module Katello
       assert_equal ::Redhat.construct_name('Red Hat Enterprise Linux'), 'RedHat'
       assert_equal ::Redhat.construct_name('My Custom Linux'), 'My_Custom_Linux'
     end
-
-    def test_assign_template
-      template = templates(:mystring2)
-      ptable = FactoryGirl.create(:ptable)
-
-      Setting.create(:name => 'katello_default_provision', :description => 'default template',
-                     :category => 'Setting::Katello', :settings_type => 'string',
-                     :default => template.name)
-
-      Setting.create(:name => 'katello_default_ptable', :description => 'default template',
-                     :category => 'Setting::Katello', :settings_type => 'string',
-                     :default => ptable.name)
-
-      os = ::Redhat.create_operating_system(@my_distro.name, '9', '0')
-      assert ::OsDefaultTemplate.where(:template_kind_id    => ::TemplateKind.find_by_name('provision').id,
-                                       :provisioning_template_id  => template.id,
-                                       :operatingsystem_id  => os.id).any?
-
-      assert os.ptables.include? ptable
-    end
   end
 end
